comfyui_morpheus_NanoBanana_Mask/requirements.txt
pillow
numpy
requests
google-genai
